requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
discord.py==1.0.0a1477+g184c430
idna==2.7
idna-ssl==1.0.1
multidict==4.3.1
pycparser==2.18
PyNaCl==1.1.2
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
websockets==5.0.1
yarl==1.2.6
